Track element position relative to center of viewport
The positoin of the element relative to the center of the viewport is now tracked. Along with visibility, now topHalf, bottomHalf, leftHalf and rightHalf state values are set. This allows you to react to the element moving into different viewport quadrants as well as visibility.

Additionally, the componentVisibilityChanged method now passes in visibilityChanged (Bool), verticalPositionChanged (Bool), horizontalPositionChanged (Bool) so that you can choose to ignore some changes over others.